+++ udata_front/theme/gouvfr/assets/js/pages/Admin/Admin.vue
@@ -100,15 +100,8 @@ onMounted(async () => {
 </script>
 
 <style lang="less">
-@import "../../../less/variables.less";
 html, body {
   height: 100%;
-}
-
-@media @dsfr-query-md {
-  .fr-sidemenu, .fr-sidemenu__inner {
-    height: 100%;
-  }
 }
 
 #app {
@@ -117,17 +110,26 @@ html, body {
   height: 100%;
 }
 
+.fr-sidemenu .fr-sidemenu__item:first-child:before, .fr-sidemenu .fr-sidemenu__item:last-child:before, .fr-sidemenu .fr-sidemenu__item:before {
+  box-shadow: none;
+}
+</style>
+<style lang="less" scoped>
+@import "../../../less/variables.less";
+
 .fr-sidemenu {
   padding: 0;
 }
 
 .fr-sidemenu .fr-sidemenu__inner {
   padding: 0;
-  box-shadow: 1px 0 0 0 var(--border-default-grey);
-  background-color: var(--background-default-grey);
+  box-shadow: 1px 0 0 0 @border-default-grey;
+  background-color: @white;
 }
 
-.fr-sidemenu .fr-sidemenu__item:first-child:before, .fr-sidemenu .fr-sidemenu__item:last-child:before, .fr-sidemenu .fr-sidemenu__item:before {
-  box-shadow: none;
+@media @dsfr-query-md {
+  .fr-sidemenu, .fr-sidemenu__inner {
+    height: 100%;
+  }
 }
 </style>